Protect routes.

+++ front-end/src/components/WorkoutDetails.js
@@ -1,21 +1,35 @@
+import { useState } from "react";
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
-
-//date fns
+import { useAuthContext } from "../hooks/useAuthContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
+  const [isDeleted, setIsDeleted] = useState(false);
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
 
     if (response.ok) {
       dispatch({ type: "DELETE_WORKOUT", payload: json });
+      setIsDeleted(true);
     }
   };
+
+  if (isDeleted) {
+    return null; // Render nothing if the workout is deleted
+  }
 
   return (
     <div className="workout-details">
